keep only used types
