+++ airtable-back/src/types/airtable.ts
@@ -59,4 +59,13 @@ export interface AirtableIngredient {
   minerals: string[];
   allergens: string[];
   unit: string;
+}
+
+export interface AirtableAllergy {
+  id: string;
+  name: string;
+  description: string;
+  keywords?: string[];
+  createdAt: string;
+  updatedAt: string;
 }
